Extract plots into partial
For the garden to be auto-generated on external contribution we need
a way to inject the generated plots into the existing site structure. So
the current plan is to use hugo itself and generate the plots with the
template and styles and then partially update the garden index on our
hoting location.